// www/js/AccessManagementUpsellV2/ts/react/legallySensitiveContent/enums/UserSetting.ts
export enum UserSetting {
  contentAgeRestriction = 'contentAgeRestriction',
  phoneNumberDiscoverability = 'phoneNumberDiscoverability',
  whoCanChatWithMe = 'whoCanChatWithMe',
  whoCanChatWithMeInApp = 'whoCanChatWithMeInApp',
  whoCanJoinMeInExperiences = 'whoCanJoinMeInExperiences',
  whoCanSeeMyInventory = 'whoCanSeeMyInventory',
  whoCanTradeWithMe = 'whoCanTradeWithMe',
  themeType = 'themeType',
  canUploadContacts = 'canUploadContacts',
  updateFriendsAboutMyActivity = 'updateFriendsAboutMyActivity',
  privateServerPrivacy = 'privateServerPrivacy',
  allowSellShareData = 'allowSellShareData',
  allowPersonalizedAdvertising = 'allowPersonalizedAdvertising',
  monthlySpendLimit = 'monthlySpendLimit',
  monthlySpendLimitCurrencyCode = 'monthlySpendLimitCurrencyCode',
  monthlySpendLimitNotificationType = 'monthlySpendLimitNotificationType',
  enablePurchases = 'enablePurchases',
  whoCanGroupChatWithMeInApp = 'whoCanGroupChatWithMeInApp',
  whoCanWhisperChatWithMeInExperiences = 'whoCanWhisperChatWithMeInExperiences',
  whoCanChatWithMeInExperiences = 'whoCanChatWithMeInExperiences',
  dailyScreenTimeLimit = 'dailyScreenTimeLimit',
  whoCanSeeMySocialNetworks = 'whoCanSeeMySocialNetworks',
  allowThirdPartyAppPermissions = 'allowThirdPartyAppPermissions',
  allowEnablePushNotifications = 'allowEnablePushNotifications',
  allowEnableEmailNotifications = 'allowEnableEmailNotifications',
  allowEnableGroupNotifications = 'allowEnableGroupNotifications',
  allowEnableExperienceNotifications = 'allowEnableExperienceNotifications',
  whoCanOneOnOnePartyWithMe = 'whoCanOneOnOnePartyWithMe',
  whoCanGroupPartyWithMe = 'whoCanGroupPartyWithMe',
  phoneNumberDiscoverabilityV2 = 'phoneNumberDiscoverabilityV2',
  whoCanPartyWithMe = 'whoCanPartyWithMe',
  whoCanUsePartyChatWithMe = 'whoCanUsePartyChatWithMe',
  whoCanUsePartyVoiceWithMe = 'whoCanUsePartyVoiceWithMe',
  allowFacialAgeEstimation = 'allowFacialAgeEstimation'
}

export default UserSetting;